% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatwalkDataset.R
\name{add_formula}
\alias{add_formula}
\title{Add a formula to a CatwalkDataset object}
\usage{
add_formula(catwalkdataset, formula)
}
\arguments{
\item{catwalkdataset}{A CatwalkDataset object to which the formula will be added}

\item{formula}{A character string specifying the formula to be added}
}
\value{
A CatwalkDataset object with the new formula added
}
\description{
Add a formula to a CatwalkDataset object
}
