% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_plots.R
\name{cal_plot}
\alias{cal_plot}
\title{Generate a single calibration plot with error bars showing 95 percent confidence intervals}
\usage{
cal_plot(df, outcome, prediction, n_bins = 10, plot_title = "", ...)
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column containing
the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column containing
the predictions.}

\item{n_bins}{Number of bins. Defaults to 10.}

\item{plot_title}{A character string containing the title for the resulting plot.}
}
\value{
A ggplot containing the calibration plot
}
\description{
Generate a single calibration plot with error bars showing 95 percent confidence intervals
}
\examples{
data(single_model_dataset)
cal_plot(single_model_dataset, outcome = 'outcomes', prediction = 'predictions', n_bins = 5)
}
