% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_models.R
\name{catwalk_anovas}
\alias{catwalk_anovas}
\title{Performs aov modeling on the dependent variables in a given dataset, using the relationship of the independent variables as dictated by the designated formula}
\usage{
catwalk_anovas(
  catwalkdataset,
  formula = 1,
  dataname = c("data_norm", "data_clean"),
  exc_vars = character(),
  cutoff = 0.05,
  make_default = F,
  allow_duplicates = F
)
}
\arguments{
\item{catwalkdataset}{A CatwalkDataset}

\item{formula}{A CatwalkDataset}

\item{dataname}{A character, or character vector, containing the name of the data within the CatwalkDataset to be used for the analysis.
If a vector of names is passed, the function will use the first name found in the CatwalkDataset}

\item{exc_vars}{A character vector of any variable names to be excluded from the analysis}

\item{cutoff}{The value at which p-values lower than this number will be considered "statistically significant"}

\item{make_default}{A boolean that will designate the output of this analysis as the default output to be reference for any future visualization or follow-up analyses}

\item{allow_duplicates}{A boolean that designates whether or not to allow the creation of two sets of models using the same data, formula, and cutoff within the same CatwalkDataset}
}
\description{
Performs aov modeling on the dependent variables in a given dataset, using the relationship of the independent variables as dictated by the designated formula
}
