% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_combine_datasets.r
\name{catwalk_combine_datasets}
\alias{catwalk_combine_datasets}
\title{Combine multiple CatwalkDataset objects into a single object}
\usage{
catwalk_combine_datasets(datasets = "list")
}
\arguments{
\item{datasets}{A list of CatwalkDataset objects to combine}
}
\value{
A new CatwalkDataset object containing merged data from all input datasets
}
\description{
Combine multiple CatwalkDataset objects into a single object
}
