% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatwalkDataset.R
\name{catwalk_get_data_from_dataset}
\alias{catwalk_get_data_from_dataset}
\title{Extract a specific dataset from a CatwalkDataset object}
\usage{
catwalk_get_data_from_dataset(
  catwalkdataset,
  dataname = c("data_norm", "data_clean")
)
}
\arguments{
\item{catwalkdataset}{A CatwalkDataset object containing the desired dataset}

\item{dataname}{A character string specifying the name of the dataset to extract (default: "data_norm")}
}
\value{
A data.frame containing the specified dataset from the CatwalkDataset object
}
\description{
Extract a specific dataset from a CatwalkDataset object
}
