% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_normalize.R
\name{catwalk_group_means}
\alias{catwalk_group_means}
\title{Collects the means for each variable in a catwalk data set for each unique group}
\usage{
catwalk_group_means(data)
}
\arguments{
\item{data}{catwalk data to measure the means of.}
}
\value{
a dataframe of means.
}
\description{
Collects the means for each variable in a catwalk data set for each unique group
}
