% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_models.R
\name{catwalk_heatmap_aov}
\alias{catwalk_heatmap_aov}
\title{Create a heatmap of significant p-values from an ANOVA model}
\usage{
catwalk_heatmap_aov(catwalkdataset, anova_results = 1)
}
\arguments{
\item{catwalkdataset}{A \code{catwalkdataset} object containing the data and ANOVA models.}

\item{anova_results}{An integer specifying the index of the ANOVA model to use for the heatmap.}
}
\value{
A \code{ggplot} object representing the heatmap.
}
\description{
Create a heatmap of significant p-values from an ANOVA model
}
