% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_pca.R
\name{catwalk_pca}
\alias{catwalk_pca}
\title{Performs PCA on a CatwalkDataset}
\usage{
catwalk_pca(catwalkdataset, ncomp = 1, dataname = c("data_norm", "data_clean"))
}
\arguments{
\item{catwalkdataset}{CatwalkDataset object to perform PCA on}

\item{ncomp}{number of principal components to keep}

\item{dataname}{name of the dataset to use for PCA}
}
\value{
the updated CatwalkDataset object containing PCA scores and raw PCA output
}
\description{
Performs PCA on a CatwalkDataset
}
