% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_models.R
\name{catwalk_process_tags}
\alias{catwalk_process_tags}
\title{Process tags for further analysis}
\usage{
catwalk_process_tags(tags, excluded_tags)
}
\arguments{
\item{tags}{A data frame with one column called "tag" containing the tags
to be processed.}

\item{excluded_tags}{A character vector of tags to exclude from analysis.}
}
\value{
A data frame with one column called "tag" containing the tags that
occur more than once, and the number of occurrences of each tag.
}
\description{
This function processes tags by filtering out excluded tags and empty tags,
counting the number of occurrences of each tag, and sorting the resulting
data frame in descending order of the number of occurrences.
}
