% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_models.R
\name{catwalk_umap}
\alias{catwalk_umap}
\title{Embed data in 2D space using Uniform Manifold Approximation and Projection (UMAP) and visualize it.}
\usage{
catwalk_umap(
  catwalkdataset,
  dataname = c("data_norm", "data_clean"),
  umap_config = umap.defaults,
  return = "plot",
  color = NULL
)
}
\arguments{
\item{catwalkdataset}{A CatWalkDataset object.}

\item{dataname}{A string specifying the name of the data set to be used in the analysis. Default is "data_norm".}

\item{umap_config}{A list of configuration parameters for the UMAP algorithm, see \code{umap.defaults} in the umap package for a full list of available parameters.}

\item{return}{A string indicating the output format. Valid options are "plot" (default), "data", or "all".}

\item{color}{A string indicating the name of the column in the data set to be used for color-coding points in the plot.}
}
\value{
If \code{return} is "plot" (default), returns a ggplot object showing the UMAP projection of the data.
If \code{return} is "data", returns a data frame with the UMAP coordinates and the columns from the input data set.
If \code{return} is "all", returns a list with the UMAP configuration parameters, the input data set, and the ggplot object.
}
\description{
Embed data in 2D space using Uniform Manifold Approximation and Projection (UMAP) and visualize it.
}
