% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_var_names.R
\name{catwalk_var_names}
\alias{catwalk_var_names}
\title{Retrieves a character vector containing the names of standard catwalk variables from catwalk version 10.}
\usage{
catwalk_var_names()
}
\value{
A character vector containing the names of standard catwalk variables from catwalk version 10.
}
\description{
Retrieves a character vector containing the names of standard catwalk variables from catwalk version 10.
}
\examples{
library(dplyr)
"Run_Duration_.s._Mean" \%in\% catwalk_var_names()

#Subset a dataset to only include standard variables
data <- read.csv(system.file("extdata",
"A111.4_Catwalk_4wks_Post_TX.csv",
package = "runway"))
data_standard <- data \%>\%
dplyr::select(catwalk_var_names())

#Warning: make sure you check what variables are being removed
#Otherwise you may accidentally remove a variable that is
#misnamed/has a name outside the standard ones
data_removed <- data \%>\% dplyr::select(!catwalk_var_names())
}
