% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_models.R
\name{catwalk_variables_to_tags}
\alias{catwalk_variables_to_tags}
\title{Convert column names to tags for catwalk functions}
\usage{
catwalk_variables_to_tags(tags, split_by)
}
\arguments{
\item{tags}{a character vector of column names to be converted to tags}

\item{split_by}{a character vector of strings to split the column names by}
}
\value{
a character vector of tags
}
\description{
Convert column names to tags for catwalk functions
}
