% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_models.R
\name{catwalk_venn}
\alias{catwalk_venn}
\title{Generate a Venn diagram from significant hits from an ANOVA model.}
\usage{
catwalk_venn(catwalkdataset, anova_results = 1, subset = NULL)
}
\arguments{
\item{catwalkdataset}{A CatWalk dataset object.}

\item{anova_results}{Index of the ANOVA model in the \code{catwalkdataset@models} list to use. Default is 1.}

\item{subset}{A function that will be applied to the hit list to determine which factors/combinations of factors to include in the graph. Default is NULL, which will include all hits.}
}
\value{
A ggvenn object representing the Venn diagram.
}
\description{
Generate a Venn diagram from significant hits from an ANOVA model.
}
