% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatwalkDataset.R
\name{CatwalkDataset}
\alias{CatwalkDataset}
\title{Create a new CatwalkDataset object from a file or dataframe}
\usage{
CatwalkDataset(
  data,
  excluded_vars = NULL,
  excluded_group = NULL,
  remove_null_vars = T,
  remove_zero_vars = T,
  group_var = "group",
  animal_var = "animal",
  experiment = "none given",
  timepoint = "none given",
  sheet = 1
)
}
\arguments{
\item{data}{A path to a spreadsheet or a dataframe containing the raw data}

\item{excluded_vars}{A character vector of variable names to exclude from the analysis}

\item{excluded_group}{A character vector of group names to exclude from the analysis}

\item{remove_null_vars}{A boolean indicating whether to remove variables containing only "-" values}

\item{remove_zero_vars}{A boolean indicating whether to remove variables containing only 0s}

\item{group_var}{A string specifying the name of the group variable in the input data}

\item{animal_var}{A string specifying the name of the animal ID variable in the input data}

\item{experiment}{A string indicating the name of the experiment (optional)}

\item{timepoint}{A string indicating the timepoint of the experiment (optional)}

\item{sheet}{An integer indicating the sheet number of the input data file (if data is a file with multiple sheets)}
}
\value{
A CatwalkDataset object containing the processed data and metadata
}
\description{
Create a new CatwalkDataset object from a file or dataframe
}
