% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_models.R
\name{fix_names}
\alias{fix_names}
\title{Fixes row names of a data frame by trimming leading/trailing whitespace}
\usage{
fix_names(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame with fixed row names
}
\description{
Fixes row names of a data frame by trimming leading/trailing whitespace
}
