% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plots.R
\name{roc_plot}
\alias{roc_plot}
\title{Generate an ROC curve plot with error bars showing 95 percent
confidence intervals}
\usage{
roc_plot(df, outcome, prediction, ci = FALSE, plot_title = "")
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column
containing the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column
containing the predictions.}

\item{ci}{Show confidence interval ribbon.
Defaults to FALSE.}

\item{plot_title}{A character string containing the title for the resulting
plot.}
}
\value{
A ggplot containing the calibration plot
}
\description{
This code builds off of code written by Vincent Guillemot found here:
\url{https://rpubs.com/vguillem/465086}.
}
\examples{
data(single_model_dataset)
roc_plot(single_model_dataset, outcome = 'outcomes', prediction = 'predictions', ci = TRUE)
}
