% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plots.R
\name{roc_plot_multi}
\alias{roc_plot_multi}
\title{Generate an ROC curve plot with error bars showing 95 percent
confidence intervals}
\usage{
roc_plot_multi(df, outcome, prediction, model, ci = FALSE, plot_title = "")
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column
containing the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column
containing the predictions.}

\item{model}{A character string containing the name of the column
containing the model label.}

\item{ci}{Show confidence interval ribbon.
Defaults to FALSE.}

\item{plot_title}{A character string containing the title for the resulting
plot.}
}
\value{
A ggplot containing the ROC plot
}
\description{
Generate an ROC curve plot with error bars showing 95 percent
confidence intervals
}
\examples{
data(multi_model_dataset)
roc_plot_multi(multi_model_dataset, outcome = 'outcomes', prediction = 'predictions', model = 'model_name', ci = TRUE)
}
