% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runway-package.R
\docType{package}
\name{runway-package}
\alias{runway}
\alias{runway-package}
\title{runway: Visualizing Prediction Model Performance}
\description{
Generate statistics and plots to calculate discrimination, calibration, and decision curves for prediction models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ML4LHS/runway}
  \item Report bugs at \url{https://github.com/ML4LHS/runway/issues}
}

}
\keyword{internal}
