% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catwalk_models.R
\name{sigmoidal_trans}
\alias{sigmoidal_trans}
\title{Define a new transformation for use with ggplot2 that maps values to the sigmoidal scale}
\usage{
sigmoidal_trans(a = 40, cutoff = 0.05)
}
\arguments{
\item{a}{The slope parameter for the sigmoidal transformation}

\item{cutoff}{The x-value at which the sigmoidal transformation transitions from flat to steep}
}
\value{
A new transformation for use with ggplot2 that maps values to the sigmoidal scale
}
\description{
This function defines a new transformation called "sigmoidal" that maps values to the sigmoidal scale.
}
