% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshperf_plots.R
\name{threshperf}
\alias{threshperf}
\title{Calculates threshold-performance data}
\usage{
threshperf(
  df,
  outcome,
  prediction,
  positive = "has_sepsis",
  thresholds = NULL,
  statistics = c("sens", "spec", "ppv", "npv"),
  prevalence = NULL
)
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column containing
the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column containing
the predictions.}

\item{statistics}{Character vector of statistics to include.
Select among \code{c("sens", "spec", "ppv", "npv", "test_pos", "test_neg", "tp_rate", "tn_rate", "fn_rate", "fp_rate", "prevalence")}.
Default is \code{c("sens", "spec", "ppv", "npv")}.}

\item{prevalence}{Specify the prevalence of outcome for case-control studies.
Default is \code{NULL} and the prevalence will be estimated from \code{data}.}
}
\value{
A data.frame containing the columns \code{.threshold}, \code{.metric},
\code{.estimator}, and \code{.estimate}
}
\description{
This function is used by the \code{\link{threshperf_plot}} function. This is based on
the \code{threshold_perf()} function from the \code{probably} package (a part of \code{tidymodels})
but expands the capability to allow for the range of metrics to include positive and
negative predictive value.
}
\examples{
data(single_model_dataset)
threshperf(single_model_dataset, outcome = 'outcomes', prediction = 'predictions')
}
