% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshperf_plots.R
\name{threshperf_plot}
\alias{threshperf_plot}
\title{Generate a threshold-performance plot for a single model}
\usage{
threshperf_plot(df, outcome, prediction, show_denom = TRUE, plot_title = "")
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column containing
the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column containing
the predictions.}

\item{show_denom}{Show the denominator (as a fraction of the maximum positives/negatives)
for the positive predictive value (PPV) and negative predictive value (NPV).}

\item{plot_title}{A character string containing the title for the resulting plot.}
}
\value{
A ggplot containing the threshold-performance plot. The 95 percent
confidence intervals are estimated using Wilson's interval from the \code{Hmisc}
\code{\link[Hmisc]{binconf}} function.
}
\description{
Generate a threshold-performance plot for a single model
}
\examples{
data(single_model_dataset)
threshperf_plot(single_model_dataset, outcome = 'outcomes', prediction = 'predictions')
}
