% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshperf_plots.R
\name{threshperf_plot_multi}
\alias{threshperf_plot_multi}
\title{Generate a threshold-performance plot for multiple models with colored/shaded 95 percent confidence intervals}
\usage{
threshperf_plot_multi(df, outcome, prediction, model, plot_title = "")
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column containing
the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column containing
the predictions.}

\item{model}{A character string containing the name of the column containing
the model names}

\item{plot_title}{A character string containing the title for the resulting plot.}
}
\value{
A ggplot containing the threshold-performance plot. The 95 percent
confidence intervals are estimated using Wilson's interval from the \code{Hmisc}
\code{\link[Hmisc]{binconf}} function.
}
\description{
Generate a threshold-performance plot for multiple models with colored/shaded 95 percent confidence intervals
}
\examples{
data(multi_model_dataset)
threshperf_plot_multi(multi_model_dataset, outcome = 'outcomes', prediction = 'predictions', model = 'model_name')
}
