% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_plots.R
\name{cal_plot}
\alias{cal_plot}
\title{A single calibration plot with error bars showing 95% confidence intervals}
\usage{
cal_plot(df, outcome_var, pred_var, n_bins = 10, plot_title = "", ...)
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome_var}{A character string containing the name of the column containing
the predictions.}

\item{pred_var}{A character string containing the name of the column containing
the predictions.}

\item{n_bins}{Number of bins. Defaults to 10.}

\item{plot_title}{A character string containing the title for the resulting plot.}
}
\value{
A ggplot containing the calibration plot
}
\description{
A single calibration plot with error bars showing 95% confidence intervals
}
