% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_plots.R
\name{cal_plot_multi}
\alias{cal_plot_multi}
\title{Multiple calibration plots with colored/shaded 95 percent confidence intervals}
\usage{
cal_plot_multi(
  df,
  outcome,
  prediction,
  model,
  n_bins = 10,
  plot_title = "",
  ...
)
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column containing
the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column containing
the predictions.}

\item{model}{A character string containing the name of the column containing
the model names}

\item{n_bins}{Number of bins. Defaults to 10.}

\item{plot_title}{A character string containing the title for the resulting plot.}
}
\value{
A ggplot containing the calibration plot
}
\description{
Multiple calibration plots with colored/shaded 95 percent confidence intervals
}
\examples{
cal_plot_multi(multi_model_dataset, outcome = 'outcomes',model = 'model_name', prediction = 'predictions', n_bins = 5)
}
