% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_plots.R
\name{cal_plot_multi}
\alias{cal_plot_multi}
\title{Generate multiple calibration plots with colored/shaded 95 percent confidence intervals}
\usage{
cal_plot_multi(
  df,
  outcome,
  prediction,
  model,
  n_bins = 10,
  show_loess = FALSE,
  plot_title = "",
  ...
)
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column containing
the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column containing
the predictions.}

\item{model}{A character string containing the name of the column containing
the model names}

\item{n_bins}{Number of bins. Defaults to 10. Set to 0 to hide binned calibration.}

\item{show_loess}{Whether to show loess smoothed calibration estimates. Defaults to FALSE.
For \code{cal_plot()}, you can display both a binned calibration plot and a loess curve. In
contrast, for \code{cal_plot_multi()}, these options are mutually exclusive. To display a loess
line, you must set \code{n_bins} to 0.}

\item{plot_title}{A character string containing the title for the resulting plot.}
}
\value{
A ggplot containing the calibration plot
}
\description{
This code builds off of code written by Darren Dahly, PhD in this blog post:
\url{https://darrendahly.github.io/post/homr/}.
}
\examples{
data(multi_model_dataset)
cal_plot_multi(multi_model_dataset, outcome = 'outcomes',model = 'model_name', prediction = 'predictions', n_bins = 5)
}
