% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshperf_plots.R
\name{threshperf}
\alias{threshperf}
\title{Calculates threshold-performance data}
\usage{
threshperf(df, outcome, prediction)
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column containing
the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column containing
the predictions.}
}
\value{
A data.frame containing the columns \code{.threshold}, \code{.metric},
\code{.estimator}, and \code{.estimate}
}
\description{
This function is used by the \code{\link{threshperf_plot}} function. This is based on
the `threshold_perf()` function from the `probably` package (a part of `tidymodels`)
but expands the capability to allow for the range of metrics to include positive and
negative predictive value.
}
\examples{
threshperf(single_model_dataset, outcome = 'outcomes', prediction = 'predictions')
}
