% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshperf_plots.R
\name{threshperf_plot}
\alias{threshperf_plot}
\title{Generate a threshold-performance plot for a single model}
\usage{
threshperf_plot(df, outcome, prediction, plot_title = "")
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column containing
the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column containing
the predictions.}

\item{plot_title}{A character string containing the title for the resulting plot.}
}
\value{
A ggplot containing the threshold-performance plot
}
\description{
Generate a threshold-performance plot for a single model
}
\examples{
threshperf_plot(single_model_dataset, outcome = 'outcomes', prediction = 'predictions')
}
