% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshperf_plots.R
\name{threshperf_plot_multi}
\alias{threshperf_plot_multi}
\title{Generate a threshold-performance plot for multiple models with colored/shaded
95 percent confidence intervals}
\usage{
threshperf_plot_multi(
  df,
  outcome,
  prediction,
  model,
  plot_title = "",
  xmin = 0,
  xmax = 1,
  pre_tp_geoms = NULL,
  pre_dist_geoms = NULL,
  post_tp_geoms = NULL,
  post_dist_geoms = NULL,
  heights = c(10, 1),
  widths = c(1, 2, 1),
  thresholds = NULL
)
}
\arguments{
\item{df}{The df as a data.frame.}

\item{outcome}{A character string containing the name of the column
containing the outcomes (expressed as 0/1s).}

\item{prediction}{A character string containing the name of the column
containing the predictions.}

\item{model}{A character string containing the name of the column containing
the model names}

\item{plot_title}{A character string containing the title for the resulting
plot.}

\item{xmin}{The minimum possible prediction. Defaults to 0.}

\item{xmax}{The maximum possible prediction. Defaults to 1.}

\item{pre_tp_geoms}{A ggplot geom_* (or a list of geoms) that should be drawn
on the threshold-performance plot prior to rendering it.}

\item{pre_dist_geoms}{A ggplot geom_* (or a list of geoms) that should be
drawn on the distribution plot prior to rendering it.}

\item{post_tp_geoms}{A ggplot geom_* (or a list of geoms) that should be
drawn on the threshold-performance plot after rendering it.}

\item{post_dist_geoms}{A ggplot geom_* (or a list of geoms) that should be
drawn on the distribution plot after rendering it.}

\item{heights}{A numeric vector of length 2 with ratio of heights of plots.
Defaults to c(10, 1).}

\item{widths}{A numeric vector of length 3 with ratio of widths of plots.
The first and third elements refer to padding. Defaults to c(1, 2, 1).}
}
\value{
A ggplot containing the threshold-performance plot. The 95 percent
confidence intervals are estimated using Wilson's interval from the
\code{Hmisc} \code{\link[Hmisc]{binconf}} function.
}
\description{
Generate a threshold-performance plot for multiple models with colored/shaded
95 percent confidence intervals
}
\examples{
data(multi_model_dataset)
threshperf_plot_multi(multi_model_dataset, outcome = 'outcomes', prediction = 'predictions', model = 'model_name')
}
